IndexedDB/submissions/Microsoft/idbobjectstore_delete5.htm

<!DOCTYPE html>
<html>
<head>
    <meta content="text/html; charset=utf-8" http-equiv="content-type" />
    <title> IDBObjectStore.delete() - removes all of the records in the range </title>
    <script type="text/javascript" src="support.js"></script>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script type="text/javascript">
        var objectStoreName = "objectstore";
        var db = null;
        
        var t = async_test();
        
        function RunTest() {
            var rqOpen = window.indexedDB.open(databaseName, databaseVersion);
            rqOpen.onupgradeneeded = t.step_func(function(event) {
                db = event.target.result;
                db.onerror = t.step_func(assert_database_error);
                var objStore = db.createObjectStore(objectStoreName);
                for(var i = 0; i < 10; i++) {
                    objStore.add({property:"data"}, i);
                }
            });
            rqOpen.onerror = t.step_func(assert_open_request_error);
            rqOpen.onsuccess = t.step_func(DeleteRecordsAndVerifyCount);
        }
        
        function DeleteRecordsAndVerifyCount() {
            var txn = db.transaction(objectStoreName, "readwrite");
            var objStore = txn.objectStore(objectStoreName);
            objStore.delete(IDBKeyRange.bound(3,6));
            var rqCount = objStore.count();
            rqCount.onsuccess = t.step_func(function(event) {
                assert_equals(event.target.result, 6, 'event.target.result');
                t.done();
            });
        }
        
        add_completion_callback(function() { if(db) db.close(); });
        
        setup(function() {
            var rqDelete = window.indexedDB.deleteDatabase(databaseName);
            rqDelete.onsuccess = t.step_func( RunTest );
        });
    </script>
</head>
<body>
    <div id="log"></div>
</body>
</html>
